+++ homeworks/brisa-rojas/18-hooks/src/components/AvatarTile/AvatarTile.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import refreshIcon from '../assets/refresh-icon.svg';
+import refreshIcon from '../../assets/refresh-icon.svg';
 import './AvatarTile.css';
 
 function AvatarTile({ avatarURL, onClick, isRefreshing }) {
